requirements.txt
pypdf
sentence-transformers
chromadb
python-dotenv
google-genai